fix issues with selling items creating a duplicate of the item in inventory, inventory order chaninging when updating an item, and price should display decimal values up to the hundredths

+++ src/components/EditItemForm.js
@@ -10,8 +10,8 @@ function EditItemForm(props) {
         props.onEditItem({
             name: event.target.name.value,
             description: event.target.description.value,
-            quantity: parseInt(event.target.quantity.value),
-            price: parseInt(event.target.price.value),
+            quantity: parseFloat(event.target.quantity.value),
+            price: parseFloat(event.target.price.value),
             id: item.id
         });
     }
@@ -20,6 +20,7 @@ function EditItemForm(props) {
         <React.Fragment>
             <ItemForms
                 formSubmissionHandler={handleEditItemFormSubmission}
+
                 buttonText="Update Item" />
         </React.Fragment>
     );
